requirements.txt

python-dotenv
pandas
scikit-learn
numpy
torch
Pillow
loguru
        